otelcolconvert: address review feedback
