+++ converter/internal/otelcolconvert/testdata/otelcol_errors/corrupt_config.yaml
@@ -1,0 +1,1 @@
+bad-key:
