+++ converter/internal/otelcolconvert/testdata/otelcol_errors/duplicate_receivers.diags
@@ -1,0 +1,1 @@
+(Critical) the configuration is unsupported because the receiver "otlp" is used across multiple pipelines with distinct names
