+++ converter/internal/otelcolconvert/testdata/otelcol_errors/duplicate_receivers.yaml
@@ -1,0 +1,21 @@
+receivers:
+  otlp:
+    protocols:
+      grpc:
+      http:
+
+exporters:
+  otlp:
+    endpoint: database:4317
+
+# A receiver may not appear in two different groups.
+service:
+  pipelines:
+    metrics: # Group <empty>
+      receivers: [otlp]
+      processors: []
+      exporters: [otlp]
+    traces/2: # Group 2
+      receivers: [otlp]
+      processors: []
+      exporters: [otlp]
