+++ converter/internal/otelcolconvert/testdata/otelcol_errors/invalid_config.diags
@@ -1,0 +1,1 @@
+(Critical) failed to validate config: no receiver configuration specified in config
